requirements.txt
numpy>=1.24.0
pillow>=10.0.0
torch>=2.0.0  # Includes MPS support for Apple Silicon
torchvision>=0.15.0
transformers>=4.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
opencv-python>=4.8.0
streamlit>=1.28.0
psutil>=5.9.0
# CuPy installation will be handled separately based on platform
tk>=0.1.0  # For tkinter
watchdog>=3.0.0  # For better file system monitoring
tqdm>=4.65.0  # For progress bars
scikit-image>=0.21.0  # For advanced image processing
matplotlib>=3.7.0  # For visualization 
jupyter>=1.0.0  # For Jupyter Notebook
faiss-cpu>=1.7.4  # For vector similarity search
# For GPU support, install faiss-gpu instead:
# faiss-gpu>=1.7.4
scikit-learn>=1.0.2
